Arg sizes
Don't rely on default argument sizes (ie. INTEGER over INTEGER(4)) as it can result in problematic translation from target language to and from fortran.

+++ sample/include/stdlib.f
@@ -2,48 +2,48 @@
 ! Standard definitions of needed functions
 ! ----------------------------------------------------------------------------
 SUBROUTINE CheckSizeInteger1(a,i)
-    INTEGER :: i
-    INTEGER :: a(:)
+    INTEGER(1) :: i
+    INTEGER(1) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
 SUBROUTINE CheckSizeInteger4(a,i)
-    INTEGER :: i
-    INTEGER :: a(:)
+    INTEGER(4) :: i
+    INTEGER(4) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
 SUBROUTINE CheckSizeInteger8(a,i)
-    INTEGER :: i
-    INTEGER :: a(:)
+    INTEGER(8) :: i
+    INTEGER(8) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
 SUBROUTINE CheckSizeReal4(a,i)
-    INTEGER :: i
+    INTEGER(4) :: i
     REAL(4) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
 SUBROUTINE CheckSizeReal8(a,i)
-    INTEGER :: i
+    INTEGER(4) :: i
     REAL(8) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
 SUBROUTINE CheckSizeReal16(a,i)
-    INTEGER :: i
+    INTEGER(4) :: i
     REAL(16) :: a(:)
 END
 
 ! ----------------------------------------------------------------------------
 
-INTEGER FUNCTION one_if(a)
+INTEGER(4) FUNCTION one_if(a)
     LOGICAL :: a
     IF (A) THEN
         one_if = 1
